refactor(utils): extract redactSecrets to its own module
